add sectors

+++ database/migrations/2020_11_03_153747_create_sectors_table.php
@@ -16,7 +16,7 @@ class CreateSectorsTable extends Migration
         Schema::create('sectors', function (Blueprint $table) {
             $table->id();
             $table->integer('size');
-            $table->float('price_per_seat');
+            $table->decimal('price_per_seat');
             $table->timestamps();
         });
     }
